Fully Implemented Dirty Checking
Dirty checking is done via a new private proptery as a byte array in the
base class. The consumer does not have to think about it. To support
this new property, I added a new SQLiteEnabledSerialization attribute to
mark up class properties, such that the programmer can tell the base
class no to place the field in the database.